brought back the circularRefHandler callback API feature from commit SHA-1: 9bd9a221c25e8f4b782d924f268b5028be32d63d, though with a slightly different API: `holder` is no longer part of the parameter set as it's the same as `objStack[index - 1]`:
* added feature (used in my JISON package test rig): `JSON5.stringify()` now has a *fourth* optional argument, which is a callback function which gets invoked when a circular reference is found in the data to stringify, which can then decide to either throw the provided exception (the classic behaviour) or produce an alternative, non-cyclic structure instead as a return value.

Interface of the callback:

    function circularRefHandler(value, index, objStack, key, err)

where `value` is the circular item, which references the already serialized entry at index `index` in the `objStack` cache. The current item is otherwise identified as the item at `key` in the `holder` object (= `objStack[index - 1]`).
`err` is a reference to the premade JSON5 exception, which may be thrown by the callback when you wish to emulate the classic `JSON5.stringify()` behaviour.