Rename to something other than JSON5, as that is incorrect and misleading
JSON5 is a specification. Anything carrying that name should adhere to the spec. In addition, the name JSON5 has special meaning, which is not true of this fork. As I stated in my suggested changes to this fork's Into:
JSON5 is designed to be valid ES5 Javascript (hence the name), [as its maintainer points out](https://github.com/json5/json5/issues/190#issuecomment-459741344):
> JSON5 indicates a specific, reliable target—ES5. JSON5 syntax will never go beyond the syntax of ES5. Sure, that means it will miss out newer features, but it also means that users don't need to worry about what version of JSONX is supported.

JSON-X doesn't maintain this guarantee. Though it includes support for ES6-style multiline string template literals,
it's not named JSON6 because it includes other syntax that is not valid ES6.